% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{occ_crosswalk}
\alias{occ_crosswalk}
\title{Occupation crosswalk dataset}
\format{
A data frame with X rows and 2 variables:
\describe{
\item{occ}{Dirty occupation names}
\item{occ_stand}{Standardized occupation names}
}
}
\source{
Your data source here
}
\usage{
occ_crosswalk
}
\description{
A dataset mapping historical "dirty" occupation names to standardized forms.
}
\keyword{datasets}
