% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occ_sector.R
\name{occ_sector}
\alias{occ_sector}
\title{Code occupations by industry}
\usage{
occ_sector(
  data,
  occ_stand,
  out_col = "standard_occupation",
  sector = TRUE,
  subsector = TRUE,
  worker_type = TRUE
)
}
\arguments{
\item{data}{A data frame containing a column with standardized occupational titles}

\item{occ_stand}{A string: the name of the column in \code{data} with standardized occupations.}

\item{out_col}{A string: name for the new column with standardized values.
Default is "standard_occupation".}

\item{sector}{A logical value: whether to code by sector. Default is TRUE.}

\item{subsector}{A logical value: whether to code by subsector. Default is TRUE.}

\item{worker_type}{A logical value: whether to code by worker type Default is TRUE.}
}
\value{
A data frame with new columns containing sector, subsector and worker type.
}
\description{
Codes occupational titles by sector, subsector, and worker type
}
\examples{
df <- data.frame(occ_stand = c("smed", "snickare", "soldat", "slöjdare", "lärling"))
occ_sector(df, "occ_stand")
}
